+++ MOVEit.MIA/Private/Invoke-MIARequest.ps1
@@ -78,6 +78,10 @@ function Invoke-MIARequest {
         # Send the request and write out the response
         Invoke-RestMethod @irmParams
     }
+    catch [System.Net.Http.HttpRequestException], [System.Net.WebException] {
+        # Format ErrorDetails which contains the JSON response from the REST API
+        $PSCmdlet.ThrowTerminatingError((Format-RestErrorDetails $PSItem))
+    }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)
     }