+++ MOVEit.MIA/Public/Disconnect-MIAServer.ps1
@@ -49,6 +49,10 @@ function Disconnect-MIAServer {
         Invoke-RestMethod @irmParams | Out-Null
         Write-Output "[$Context]: Disconnected from MOVEit Automation server"
     }
+    catch [System.Net.Http.HttpRequestException], [System.Net.WebException] {
+        # Format ErrorDetails which contains the JSON response from the REST API
+        $PSCmdlet.ThrowTerminatingError((Format-RestErrorDetails $PSItem))
+    }
     catch {
         $PSCmdlet.ThrowTerminatingError($PSItem)
     }